coords parser changed
refactoring tests
gaus-krueger "-" allowed
maidenhead parser invalid coords check changed
utm parser bugfix
