+++ lib/logic/tools/coords/converter/gauss_krueger.dart
@@ -126,7 +126,7 @@ String latLonToGaussKruegerString(LatLng coords, int gaussKruegerCode, Ellipsoid
 }
 
 LatLng parseGaussKrueger(String input, Ellipsoid ells, {gaussKruegerCode: 1}) {
-  RegExp regExp = RegExp(r'^\s*([0-9\.]+)(\s*\,\s*|\s+)([0-9\.]+)\s*$');
+  RegExp regExp = RegExp(r'^\s*([\-0-9\.]+)(\s*\,\s*|\s+)([\-0-9\.]+)\s*$');
   var matches = regExp.allMatches(input);
   var _eastingString = '';
   var _northingString = '';
@@ -137,7 +137,7 @@ LatLng parseGaussKrueger(String input, Ellipsoid ells, {gaussKruegerCode: 1}) {
     _northingString = match.group(3);
   }
   if (matches.length == 0) {
-    regExp = RegExp(r'^\s*(R|r)\:?\s*([0-9\.]+)(\s*\,?\s*)(H|h)\:?\s*([0-9\.]+)\s*$');
+    regExp = RegExp(r'^\s*(R|r)\:?\s*([\-0-9\.]+)(\s*\,?\s*)(H|h)\:?\s*([\-0-9\.]+)\s*$');
     matches = regExp.allMatches(input);
     if (matches.length > 0) {
       var match = matches.elementAt(0);
