+++ lib/logic/tools/coords/converter/maidenhead.dart
@@ -15,18 +15,24 @@ LatLng maidenheadToLatLon (String maidenhead) {
   try {
     for (int i = 0; i < maidenhead.length; i += 2) {
       if (res == 1) {
+        if (!RegExp(r'[A-Z]{2}').hasMatch(maidenhead.substring(0, 2)))
+          return null;
         lon = ((alphabet_AZ[maidenhead[0]] - 1) * 20).toDouble();
         lat = ((alphabet_AZ[maidenhead[1]] - 1) * 10).toDouble();
         res = 2;
       } else if (res % 2 == 1)  {
         reslon /= 24;
         reslat /= 24;
+        if (!RegExp(r'[A-Z]{2}').hasMatch(maidenhead.substring(i, i+2)))
+          return null;
         lon += (alphabet_AZ[maidenhead[i]] - 1).toDouble() * reslon;
         lat += (alphabet_AZ[maidenhead[i + 1]] - 1).toDouble() * reslat;
         ++res;
       } else {
         reslon /= 10;
         reslat /= 10;
+        if (!RegExp(r'[0-9]{2}').hasMatch(maidenhead.substring(i, i+2)))
+          return null;
         lon += int.tryParse(maidenhead[i]) * reslon;
         lat += int.tryParse(maidenhead[i + 1]) * reslat;
         ++res;
@@ -35,8 +41,6 @@ LatLng maidenheadToLatLon (String maidenhead) {
   } catch(e) {
     return null;
   }
-  if (lat == 0.0 && lon == 0.0)
-    return null;
 
   lon -= 180;
   lat -= 90;
